+++ index.js
@@ -1,28 +1,23 @@
 const goodbye = require('graceful-goodbye')
-const Replicator = require('hypercore/lib/replicator')
+const Corestore = require('corestore')
+const Hyperswarm = require('hyperswarm')
+const DHT = require('hyperdht')
+const SimpleSeeder = require('simple-seeder/lib/simple-seeder')
+const debounceify = require('debounceify')
 
 const instrument = require('./lib/instrument')
 const getSeederInfo = require('./lib/seeder-info')
-const setupSeeder = require('./lib/seeder')
 
 module.exports = async function runSeeder (logger, config) {
-  if (config.trace) {
-    logger.warn('Applying monkey patches to improve tracing')
-    // TODO: clean up flow
-    // Metrics should be setup before the seeder starts, then
-    // the monkey patching can live in the metrics logic itself
-    applyTracerMonkeyPatches()
-  }
+  const { swarm, store } = await setupSwarmAndStore(config) // TODO: would be cleaner as a sync function
+  const tracker = new SimpleSeeder(store, swarm)
 
-  const tracker = await setupSeeder(config)
   goodbye(async () => {
     logger.info('Exiting simple seeder')
     await tracker.destroy()
-    await tracker.swarm.destroy()
+    await swarm.destroy()
     logger.info('Destroyed swarm and tracker')
   }, 10)
-
-  logger.info('Setup simple seeder')
 
   if (config.instrument) {
     const server = await instrument(tracker, logger, config)
@@ -35,6 +30,9 @@ module.exports = async function runSeeder (logger, config) {
     logger.info('Instrumented the simple seeder')
   }
 
+  await startSeeder(tracker, config.seedListKey)
+  logger.info('Setup simple seeder')
+
   if (config.sLogInterval) {
     setInterval(
       () => { logger.info(getSeederInfo(tracker)) },
@@ -44,10 +42,34 @@ module.exports = async function runSeeder (logger, config) {
   }
 }
 
-function applyTracerMonkeyPatches () {
-  const originalRequestRangeBlock = Replicator.Peer.prototype._requestRangeBlock
-  Replicator.Peer.prototype._requestRangeBlock = function _requestRangeBlockMonkeyPatch (...args) {
-    this.tracer.trace('_requestRangeBlock')
-    return originalRequestRangeBlock.call(this, ...args)
+async function setupSwarmAndStore ({ corestoreLoc, port, maxPeers }) {
+  const store = new Corestore(corestoreLoc)
+
+  const keyPair = await store.createKeyPair('simple-seeder-swarm')
+  // Assume that if a DHT port was set, it's not firewalled
+  const firewalled = port == null || port === 0
+  const dht = new DHT({ port, firewalled })
+  const swarm = new Hyperswarm({ keyPair, dht, maxPeers })
+
+  swarm.on('connection', (socket) => {
+    store.replicate(socket)
+  })
+
+  return { swarm, store }
+}
+
+async function startSeeder (tracker, seedListKey) {
+  await tracker.add(seedListKey, { type: 'list', description: 'Simple Seeder main list' })
+
+  // TODO: figure out why this logic is outside simple-seeder
+  const lists = tracker.filter(r => r.type === 'list')
+  if (lists[0]) {
+    const info = lists[0]
+    const bound = tracker.update.bind(tracker, info, info.instance)
+    const debounced = debounceify(bound)
+    info.instance.core.on('append', debounced)
+    await debounced()
   }
+
+  return tracker
 }
