+++ lib/instrument.js
@@ -6,6 +6,7 @@ const setupRepl = require('repl-swarm')
 const idEncoding = require('hypercore-id-encoding')
 const InstrumentedCorestore = require('@hdegroote/instrumented-corestore')
 const Hypertrace = require('hypertrace')
+const Replicator = require('hypercore/lib/replicator')
 
 const { version: PACKAGE_VERSION, name: PACKAGE_NAME } = require('../package.json')
 
@@ -17,16 +18,23 @@ const PEER_METRICS = [
 ]
 
 module.exports = async function instrument (tracker, logger, {
-  repl, instrumentPort, instrumentHost, detailedMetrics, seedListKey
+  repl, instrumentPort, instrumentHost, detailedMetrics, seedListKey, trace
 }) {
-  const counters = new Map()
-  Hypertrace.setTraceFunction(({ id, object }) => {
-    if (object.className !== 'Peer') return
+  if (trace) {
+    logger.warn('Applying monkey patches to improve tracing')
+    applyTracerMonkeyPatches()
 
-    const uid = `${object.className}-${id}`
-    const prev = counters.get(uid) || 0
-    counters.set(uid, prev + 1)
-  })
+    const counters = new Map()
+    Hypertrace.setTraceFunction(({ id, object }) => {
+      if (object.className !== 'Peer') return
+
+      const uid = `${object.className}-${id}`
+      const prev = counters.get(uid) || 0
+      counters.set(uid, prev + 1)
+    })
+
+    registerTracingMetrics(counters)
+  }
 
   const server = fastify({ logger })
 
@@ -53,7 +61,6 @@ module.exports = async function instrument (tracker, logger, {
   promClient.collectDefaultMetrics()
   instrumentedSwarm.registerPrometheusMetrics(promClient)
   registerPackageVersion()
-  registerTracingMetrics(counters)
 
   server.get('/metrics', { logLevel: 'warn' }, async function (req, reply) {
     const metrics = await promClient.register.metrics()
@@ -127,4 +134,12 @@ function registerTracingMetrics (counters) {
   }
 
   return metrics
+}
+
+function applyTracerMonkeyPatches () {
+  const originalRequestRangeBlock = Replicator.Peer.prototype._requestRangeBlock
+  Replicator.Peer.prototype._requestRangeBlock = function _requestRangeBlockMonkeyPatch (...args) {
+    this.tracer.trace('_requestRangeBlock')
+    return originalRequestRangeBlock.call(this, ...args)
+  }
 }
